Break apart module into separate files
