+++ src/euler_library/problems/large_sum.rs
@@ -1,0 +1,3 @@
+pub fn compute() -> String {
+    "Hello World".to_string()
+}
